+++ src/nti/testing/tests/test_main.py
@@ -13,6 +13,20 @@ class TestImport(unittest.TestCase):
         for name in ('base', 'layers', 'matchers', 'time'):
             __import__('nti.testing.' + name)
 
+    def test_mock_direct_import(self):
+        from nti.testing.mock import Mock
+        from nti.testing import mock
+        self.assertIs(mock.Mock, Mock)
+
+    def test_mock_module_identity(self):
+        from nti.testing import mock
+        try:
+            from unittest import mock as stdmock
+        except ImportError: # pragma: no cover
+            # Python 2
+            import mock as stdmock
+
+        self.assertIs(mock, stdmock)
 
 def test_suite():
     here = os.path.dirname(__file__)